backend/requirements.txt
flask
requests
python-docx
python-dotenv
# Add any other dependencies here, one per line
